Extract taints-and-tolerations.go from infra_test.go
